+++ packages/gatsby-plugin-manifest/src/gatsby-node.js
@@ -35,7 +35,7 @@ exports.onPostBootstrap = (args, pluginOptions) =>
     // Determine destination path for icons.
     const iconPath = path.join(
       `public`,
-      manifest.icons[0].src.substring(0, manifest.icons[0].src.lastIndexOf(`/`))
+      path.dirname(manifest.icons[0].src)
     )
 
     //create destination directory if it doesn't exist