+++ packages/gatsby-transformer-screenshot/src/gatsby-node.js
@@ -12,7 +12,7 @@ const screenshotQueue = new Queue(
       .then(r => cb(null, r))
       .catch(e => cb(e))
   },
-  { concurrent: LAMBDA_CONCURRENCY_LIMIT, maxRetries: 10, retryDelay: 1000 }
+  { concurrent: LAMBDA_CONCURRENCY_LIMIT, maxRetries: 3, retryDelay: 1000 }
 )
 
 const createContentDigest = obj =>
@@ -68,42 +68,47 @@ exports.onPreBootstrap = (
   })
 }
 
-exports.onCreateNode = async ({
-  node,
-  actions,
-  store,
-  cache,
-  createNodeId,
-}) => {
+exports.onCreateNode = async (
+  { node, actions, store, cache, createNodeId },
+  pluginOptions
+) => {
   const { createNode, createParentChildLink } = actions
 
-  // We only care about parsed sites.yaml files with a url field
-  if (node.internal.type !== `SitesYaml` || !node.url) {
+  /*
+   * Check if node is of a type we care about, and has a url field
+   * (originally only checked sites.yml, hence including by default)
+   */
+  const validNodeTypes = [`SitesYaml`].concat(pluginOptions.nodeTypes || [])
+  if (!validNodeTypes.includes(node.internal.type) || !node.url) {
     return
   }
 
-  const screenshotNode = await new Promise((resolve, reject) => {
-    screenshotQueue
-      .push({
-        url: node.url,
-        parent: node.id,
-        store,
-        cache,
-        createNode,
-        createNodeId,
-      })
-      .on(`finish`, r => {
-        resolve(r)
-      })
-      .on(`failed`, e => {
-        reject(e)
-      })
-  })
+  try {
+    const screenshotNode = await new Promise((resolve, reject) => {
+      screenshotQueue
+        .push({
+          url: node.url,
+          parent: node.id,
+          store,
+          cache,
+          createNode,
+          createNodeId,
+        })
+        .on(`finish`, r => {
+          resolve(r)
+        })
+        .on(`failed`, e => {
+          reject(e)
+        })
+    })
 
-  createParentChildLink({
-    parent: node,
-    child: screenshotNode,
-  })
+    createParentChildLink({
+      parent: node,
+      child: screenshotNode,
+    })
+  } catch (e) {
+    return
+  }
 }
 
 const createScreenshotNode = async ({