+++ packages/gatsby/cache-dir/navigation.js
@@ -1,8 +1,13 @@
+import React from "react"
+import PropTypes from "prop-types"
 import loader, { setApiRunnerForLoader } from "./loader"
 import redirects from "./redirects.json"
 import { apiRunner } from "./api-runner-browser"
 import emitter from "./emitter"
-import { resolveRouteChangeListeners } from "./wait-for-route-change"
+import {
+  resolveRouteChangePromise,
+  resetRouteChangePromise,
+} from "./wait-for-route-change"
 import { navigate as reachNavigate } from "@reach/router"
 import parsePath from "./parse-path"
 import loadDirectlyOr404 from "./load-directly-or-404"
@@ -39,10 +44,11 @@ const onPreRouteUpdate = location => {
     apiRunner(`onPreRouteUpdate`, { location })
   }
 }
+
 const onRouteUpdate = location => {
   if (!maybeRedirect(location.pathname)) {
     apiRunner(`onRouteUpdate`, { location })
-    resolveRouteChangeListeners()
+    resolveRouteChangePromise()
 
     // Temp hack while awaiting https://github.com/reach/router/issues/119
     window.__navigatingToLink = false
@@ -71,6 +77,8 @@ const navigate = (to, options = {}) => {
     return
   }
 
+  resetRouteChangePromise()
+
   // Start a timer to wait for a second before transitioning and showing a
   // loader in case resources aren't around yet.
   const timeoutId = setTimeout(() => {
@@ -84,18 +92,20 @@ const navigate = (to, options = {}) => {
     if (!pageResources && process.env.NODE_ENV === `production`) {
       loader.getResourcesForPathname(`/404.html`).then(resources => {
         clearTimeout(timeoutId)
-        onPreRouteUpdate(window.location)
-        loadDirectlyOr404(resources, to).then(() =>
-          reachNavigate(to, options).then(() => onRouteUpdate(window.location))
-        )
+        loadDirectlyOr404(resources, to).then(() => reachNavigate(to, options))
       })
     } else {
-      onPreRouteUpdate(window.location)
-      reachNavigate(to, options).then(() => onRouteUpdate(window.location))
+      reachNavigate(to, options)
       clearTimeout(timeoutId)
     }
   })
 }
+
+// reset route change promise after going back / forward
+// in history (when not using Gatsby navigation)
+window.addEventListener(`popstate`, () => {
+  resetRouteChangePromise()
+})
 
 function shouldUpdateScroll(prevRouterProps, { location: { pathname } }) {
   const results = apiRunner(`shouldUpdateScroll`, {
@@ -131,4 +141,39 @@ function init() {
   maybeRedirect(window.location.pathname)
 }
 
-export { init, shouldUpdateScroll, onRouteUpdate, onPreRouteUpdate }
+// Fire on(Pre)RouteUpdate APIs
+class RouteUpdates extends React.Component {
+  constructor(props) {
+    super(props)
+    onPreRouteUpdate(props.location)
+  }
+
+  componentDidMount() {
+    onRouteUpdate(this.props.location)
+  }
+
+  componentDidUpdate(prevProps, prevState, shouldFireRouteUpdate) {
+    if (shouldFireRouteUpdate) {
+      onRouteUpdate(this.props.location)
+    }
+  }
+
+  getSnapshotBeforeUpdate(prevProps) {
+    if (this.props.location.pathname !== prevProps.location.pathname) {
+      onPreRouteUpdate(this.props.location)
+      return true
+    }
+
+    return false
+  }
+
+  render() {
+    return this.props.children
+  }
+}
+
+RouteUpdates.propTypes = {
+  location: PropTypes.object.isRequired,
+}
+
+export { init, shouldUpdateScroll, RouteUpdates }