+++ www/gatsby-node.js
@@ -48,6 +48,12 @@ exports.createPages = ({ graphql, actions }) => {
     toPath: `/docs/sourcing-from-netlify-cms`,
   })
 
+  createRedirect({
+    fromPath: `/starter-showcase/`, // Moved "Starter Showcase" index page from /starter-showcase to /starters
+    toPath: `/starters/`,
+    isPermanent: true,
+  })
+
   return new Promise((resolve, reject) => {
     const docsTemplate = path.resolve(`src/templates/template-docs-markdown.js`)
     const blogPostTemplate = path.resolve(`src/templates/template-blog-post.js`)
@@ -64,6 +70,9 @@ exports.createPages = ({ graphql, actions }) => {
     )
     const showcaseTemplate = path.resolve(
       `src/templates/template-showcase-details.js`
+    )
+    const creatorPageTemplate = path.resolve(
+      `src/templates/template-creator-details.js`
     )
 
     createRedirect({
@@ -111,6 +120,15 @@ exports.createPages = ({ graphql, actions }) => {
             }
           }
           allAuthorYaml {
+            edges {
+              node {
+                fields {
+                  slug
+                }
+              }
+            }
+          }
+          allCreatorsYaml {
             edges {
               node {
                 fields {
@@ -169,7 +187,7 @@ exports.createPages = ({ graphql, actions }) => {
 
       Array.from({ length: numPages }).forEach((_, i) => {
         createPage({
-          path: i === 0 ? `/blog` : `/blog/${i + 1}`,
+          path: i === 0 ? `/blog` : `/blog/page/${i + 1}`,
           component: slash(blogListTemplate),
           context: {
             limit: postsPerPage,
@@ -203,7 +221,7 @@ exports.createPages = ({ graphql, actions }) => {
 
       _.uniq(_.flatten(tagLists)).forEach(tag => {
         createPage({
-          path: `/blog/tags/${_.kebabCase(tag)}/`,
+          path: `/blog/tags/${_.kebabCase(tag.toLowerCase())}/`,
           component: tagTemplate,
           context: {
             tag,
@@ -238,6 +256,18 @@ exports.createPages = ({ graphql, actions }) => {
         createPage({
           path: `${edge.node.fields.slug}`,
           component: slash(contributorPageTemplate),
+          context: {
+            slug: edge.node.fields.slug,
+          },
+        })
+      })
+
+      result.data.allCreatorsYaml.edges.forEach(edge => {
+        if (!edge.node.fields) return
+        if (!edge.node.fields.slug) return
+        createPage({
+          path: `${edge.node.fields.slug}`,
+          component: slash(creatorPageTemplate),
           context: {
             slug: edge.node.fields.slug,
           },
@@ -372,6 +402,28 @@ exports.onCreateNode = ({ node, actions, getNode, getNodes }) => {
     slug = `/showcase/${slugify(cleaned)}`
     createNodeField({ node, name: `slug`, value: slug })
   }
+
+  // Community/Creators Pages
+  else if (node.internal.type === `CreatorsYaml`) {
+    const validTypes = {
+      individual: `people`,
+      agency: `agencies`,
+      company: `companies`,
+    }
+
+    if (!validTypes[node.type]) {
+      throw new Error(
+        `Creators must have a type of “individual”, “agency”, or “company”, but invalid type “${
+          node.type
+        }” was provided for ${node.name}.`
+      )
+    }
+    slug = `/community/${validTypes[node.type]}/${slugify(node.name, {
+      lower: true,
+    })}`
+    createNodeField({ node, name: `slug`, value: slug })
+  }
+  // end Community/Creators Pages
 }
 
 exports.onPostBuild = () => {