+++ www/src/components/layout.js
@@ -1,7 +1,6 @@
 import React from "react"
 import Modal from "react-modal"
 import Helmet from "react-helmet"
-import { OutboundLink } from "gatsby-plugin-google-analytics"
 import MdClose from "react-icons/lib/md/close"
 import { navigate, PageRenderer } from "gatsby"
 import presets, { colors } from "../utils/presets"
@@ -9,7 +8,6 @@ import Banner from "../components/banner"
 import Navigation from "../components/navigation"
 import MobileNavigation from "../components/navigation-mobile"
 import PageWithSidebar from "../components/page-with-sidebar"
-import "../css/prism-coy.css"
 
 import mousetrap from "mousetrap"
 
@@ -58,12 +56,7 @@ class DefaultLayout extends React.Component {
   }
 
   render() {
-    const {
-      location = {
-        pathname: `/starter-showcase`,
-      },
-    } = this.props // location will be undefined if on 'starter-showcase'
-    const isHomepage = location.pathname === `/`
+    const isHomepage = this.props.location.pathname === `/`
 
     // SEE: template-docs-markdown for why this.props.isSidebarDisabled is here
     const isSidebarDisabled =
@@ -165,10 +158,10 @@ class DefaultLayout extends React.Component {
           <html lang="en" />
         </Helmet>
         <Banner background={isHomepage ? `#402060` : false}>
-          These are the docs for v2 beta.
+          These are the docs for v2.
           {` `}
-          <OutboundLink
-            href="https://gatsbyjs.org/"
+          <a
+            href="https://v1.gatsbyjs.org/"
             css={{
               color: `#fff`,
             }}
@@ -185,7 +178,7 @@ class DefaultLayout extends React.Component {
               {` `}
               instead
             </span>
-          </OutboundLink>
+          </a>
           .
         </Banner>
         <Navigation pathname={this.props.location.pathname} />