+++ www/src/components/showcase-details.js
@@ -164,7 +164,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                     ...styles.prevNextImage,
                   }}
                   backgroundColor
-                  resolutions={{
+                  fixed={{
                     srcSet: ``,
                     src:
                       nextSite.childScreenshot.screenshotFile.childImageSharp
@@ -216,7 +216,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                     ...styles.prevNextImage,
                   }}
                   backgroundColor
-                  resolutions={{
+                  fixed={{
                     srcSet: ``,
                     src:
                       previousSite.childScreenshot.screenshotFile
@@ -265,14 +265,14 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                 <title>{data.sitesYaml.title}</title>
                 <meta
                   name="og:image"
-                  content={`https://next.gatsbyjs.org${
+                  content={`https://www.gatsbyjs.org${
                     data.sitesYaml.childScreenshot.screenshotFile
                       .childImageSharp.resize.src
                   }`}
                 />
                 <meta
                   name="twitter:image"
-                  content={`https://next.gatsbyjs.org${
+                  content={`https://www.gatsbyjs.org${
                     data.sitesYaml.childScreenshot.screenshotFile
                       .childImageSharp.resize.src
                   }`}
@@ -477,7 +477,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                   <ShareMenu
                     url={data.sitesYaml.main_url}
                     title={data.sitesYaml.title}
-                    image={`https://next.gatsbyjs.org${
+                    image={`https://www.gatsbyjs.org${
                       data.sitesYaml.childScreenshot.screenshotFile
                         .childImageSharp.resize.src
                     }`}
@@ -485,9 +485,9 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                 </div>
                 <Img
                   key={data.sitesYaml.id}
-                  sizes={
+                  fluid={
                     data.sitesYaml.childScreenshot.screenshotFile
-                      .childImageSharp.sizes
+                      .childImageSharp.fluid
                   }
                   alt={`Screenshot of ${data.sitesYaml.title}`}
                   css={{