+++ www/src/views/starter-showcase/showcase-list.js
@@ -1,14 +1,14 @@
-import React /* { Component } */ from "react"
-import { Link } from "gatsby"
-import Img from "gatsby-image"
-import FaExtLink from "react-icons/lib/fa/external-link"
-import FaGithub from "react-icons/lib/fa/github"
-import FaClipboard from "react-icons/lib/fa/clipboard"
+import React from "react"
+import LaunchDemoIcon from "react-icons/lib/md/launch"
+import GithubIcon from "react-icons/lib/go/mark-github"
+import CopyToClipboardIcon from "react-icons/lib/go/clippy"
 import MdStar from "react-icons/lib/md/star"
-import { /* rhythm, */ scale, rhythm, options } from "../../utils/typography"
-import presets, { colors } from "../../utils/presets"
+import { options } from "../../utils/typography"
+import { colors } from "../../utils/presets"
 import copyToClipboard from "../../utils/copy-to-clipboard"
 import styles from "../shared/styles"
+import ThumbnailLink from "../shared/thumbnail"
+import EmptyGridItems from "../shared/empty-grid-items"
 
 const ShowcaseList = ({ urlState, items, imgs, count, sortRecent }) => {
   if (!items.length) {
@@ -50,13 +50,7 @@ const ShowcaseList = ({ urlState, items, imgs, count, sortRecent }) => {
     <div
       css={{
         fontFamily: options.headerFontFamily.join(`,`),
-        display: `flex`,
-        flexWrap: `wrap`,
-        padding: rhythm(3 / 4),
-        justifyContent: `center`,
-        [presets.Desktop]: {
-          justifyContent: `flex-start`,
-        },
+        ...styles.showcaseList,
       }}
     >
       {items.map(({ node }) => {
@@ -75,6 +69,7 @@ const ShowcaseList = ({ urlState, items, imgs, count, sortRecent }) => {
         const minorVersion = match ? match[0] : gatsbyVersion // default to version if no match
         const isGatsbyVersionWarning = !/(2..+|next|latest)/g.test(minorVersion) // either 2.x or next or latest
         const imgsharp = imgsFilter(imgs, stub)
+
         const repo = githubData.repoMetadata
         const { pushed_at } = repo
         return (
@@ -82,107 +77,51 @@ const ShowcaseList = ({ urlState, items, imgs, count, sortRecent }) => {
             <div
               key={node.id}
               css={{
-                margin: rhythm(3 / 4),
-                width: 280,
+                ...styles.showcaseItem,
               }}
               {...styles.withTitleHover}
             >
-              <Link
-                to={`/starters/${stub}`}
-                state={{ isModal: true }}
-                css={{
-                  "&&": {
-                    borderBottom: `none`,
-                    boxShadow: `none`,
-                    transition: `all ${presets.animation.speedDefault} ${
-                      presets.animation.curveDefault
-                    }`,
-                    "&:hover": {
-                      ...styles.screenshotHover,
-                      // manual patch for hover
-                      boxShadow: `0 8px 20px rgba(157,124,191,0.5)`,
-                      transform: `translateY(-3px)`,
-                    },
-                  },
-                }}
-              >
-                <div
-                  className="gatsby-image-wrapper"
-                  css={{
-                    transition: `all ${presets.animation.speedDefault} ${
-                      presets.animation.curveDefault
-                    }`,
-                  }}
-                >
-                  {imgsharp ? (
-                    <Img
-                      fluid={imgsharp.childImageSharp.fluid}
-                      alt={`Screenshot of ${imgsharp.name}`}
-                      css={{
-                        ...styles.screenshot,
-                        marginBottom: 0,
-                      }}
-                    />
-                  ) : (
-                    <div
-                      css={{
-                        // height: 230,
-                        width: 320,
-                        backgroundColor: `#d999e7`,
-                      }}
-                    >
-                      missing
-                    </div>
-                  )}
-                </div>
-              </Link>
+              <ThumbnailLink
+                slug={`/starters/${stub}`}
+                image={imgsharp}
+                title={imgsharp.name}
+              />
               <div
                 css={{
                   ...styles.meta,
-                  marginTop: rhythm(options.blockMarginBottom),
                 }}
               >
                 <div css={{ display: `flex`, justifyContent: `space-between` }}>
                   <span css={{ color: colors.gray.dark }}>
                     {repo.owner && repo.owner.login} /
                   </span>
-                  <span
-                    css={{
-                      "> a": {
-                        paddingLeft: 5,
-                        "&:hover": {
-                          background: `none`,
-                          color: colors.gatsby,
-                        },
-                      },
-                    }}
-                  >
+                  <span>
                     <a
                       href="#copy-to-clipboard"
                       onClick={() =>
                         copyToClipboard(`https://github.com/${githubFullName}`)
                       }
-                      css={{ ...styles.noLinkUnderline }}
+                      css={{ ...styles.shortcutIcon }}
                     >
-                      <FaClipboard />
-                      {` `}
+                      <CopyToClipboardIcon />
                     </a>
+                    {` `}
                     <a
                       href={node.frontmatter.demo}
                       target="_blank"
                       rel="noopener noreferrer"
-                      css={{ ...styles.noLinkUnderline }}
+                      css={{ ...styles.shortcutIcon }}
                     >
-                      <FaExtLink />
-                      {` `}
+                      <LaunchDemoIcon />
                     </a>
+                    {` `}
                     <a
                       href={`https://github.com/${githubFullName}`}
                       target="_blank"
                       rel="noopener noreferrer"
-                      css={{ ...styles.noLinkUnderline }}
+                      css={{ ...styles.shortcutIcon }}
                     >
-                      <FaGithub />
+                      <GithubIcon />
                       {` `}
                     </a>
                   </span>
@@ -226,6 +165,7 @@ const ShowcaseList = ({ urlState, items, imgs, count, sortRecent }) => {
           )
         )
       })}
+      {items.length && <EmptyGridItems styles={styles.showcaseItem} />}
     </div>
   )
 }