+++ packages/gatsby-image/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.9"></a>
+
+## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.8...gatsby-image@2.0.9) (2018-09-24)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.8"></a>
+
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.7...gatsby-image@2.0.8) (2018-09-21)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.7"></a>
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.6...gatsby-image@2.0.7) (2018-09-20)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.6"></a>
+
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.5...gatsby-image@2.0.6) (2018-09-19)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.4...gatsby-image@2.0.5) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.4"></a>
+
+# [2.0.0-rc.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.3...gatsby-image@2.0.0-rc.4) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.3"></a>
+
+# [2.0.0-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.2...gatsby-image@2.0.0-rc.3) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.2"></a>
+
+# [2.0.0-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.1...gatsby-image@2.0.0-rc.2) (2018-09-12)
+
+**Note:** Version bump only for package gatsby-image
+
 <a name="2.0.0-rc.1"></a>
 
 # [2.0.0-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.0...gatsby-image@2.0.0-rc.1) (2018-08-29)